+++ AppleFramework/Screens/FrameworkGridView/FrameworkGridView.swift
@@ -37,21 +37,3 @@ struct FrameworkGridView: View {
         .preferredColorScheme(.dark)
 }
 
-struct FrameworkTitleView: View {
-    let framework: Framework
-    
-    var body: some View {
-        VStack {
-            Image(framework.imageName)
-                .resizable()
-                .frame(width: 90, height: 90)
-            
-            Text(framework.name)
-                .font(.title2)
-                .fontWeight(.semibold)
-                .scaledToFit()
-                .minimumScaleFactor(0.6)
-        }
-        .padding()
-    }
-}
