+++ AppleFramework/Screens/FrameworkGridView/FrameworkGridViewModel.swift
@@ -8,20 +8,13 @@
 import SwiftUI
 
 final class FrameworkGridViewModel: ObservableObject {
-    
-    let columns: [GridItem] = [ GridItem(.flexible()),
-                                GridItem(.flexible()),
-                                GridItem(.flexible()) ]
-    
     var selectedFramework: Framework? {
-        didSet {
-            isShowingDetailView = true
-        }
+        didSet { isShowingDetailView = true }
     }
     
     @Published var isShowingDetailView = false
     
-    
-
-    
+    let columns: [GridItem] = [ GridItem(.flexible()),
+                                GridItem(.flexible()),
+                                GridItem(.flexible()) ]
 }
