+++ AppleFramework/Views/View/FrameworkTitleView.swift
@@ -1,0 +1,32 @@
+//
+//  FrameworkTitleView.swift
+//  AppleFramework
+//
+//  Created by Christian Manzaraz on 31/07/2024.
+//
+
+import SwiftUI
+
+struct FrameworkTitleView: View {
+    let framework: Framework
+    
+    var body: some View {
+        VStack {
+            Image(framework.imageName)
+                .resizable()
+                .frame(width: 90, height: 90)
+            
+            Text(framework.name)
+                .font(.title2)
+                .fontWeight(.semibold)
+                .scaledToFit()
+                .minimumScaleFactor(0.6)
+        }
+        .padding()
+    }
+}
+
+
+#Preview {
+    FrameworkTitleView(framework: MockData.sampleFramework)
+}
